requirements.txt
appdirs==1.4.4
black==21.5b0
click==7.1.2
colorama==0.4.4
Flask==2.0.0
flask-marshmallow==0.14.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
mando==0.6.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.0
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
numpy==1.20.3
pandas==1.2.4
parso==0.8.2
pathspec==0.8.1
plotly==4.14.3
pycodestyle==2.7.0
pydocstyle==6.0.0
pyflakes==2.3.1
pylava==0.3.0
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
radon==4.5.0
regex==2021.4.4
retrying==1.3.3
six==1.15.0
snowballstemmer==2.1.0
SQLAlchemy==1.4.15
toml==0.10.2
Werkzeug==2.0.1
